% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMisscleavedPeptides.R
\docType{methods}
\name{addMisscleavedPeptides}
\alias{addMisscleavedPeptides}
\alias{addMisscleavedPeptides,SilacProteinExperiment-method}
\alias{addMisscleavedPeptides,SilacPeptideExperiment-method}
\alias{addMisscleavedPeptides,SilacProteomicsExperiment-method}
\title{Add miss-cleaved peptide expression levels for old isotope recycling
estimation}
\usage{
addMisscleavedPeptides(x, ...)

\S4method{addMisscleavedPeptides}{SilacProteinExperiment}(x, newdata,
  modCol, dataCols, idColPept)

\S4method{addMisscleavedPeptides}{SilacPeptideExperiment}(x, newdata,
  modCol, dataCols, idColPept)

\S4method{addMisscleavedPeptides}{SilacProteomicsExperiment}(x, newdata,
  modCol, dataCols, idColPept)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or
\code{SilacProteomicsExperiment}.}

\item{...}{Unused.}

\item{newdata}{a \code{data.frame} containing the data described in the
description.}

\item{modCol}{\code{character} or \code{numeric} indicating which column of
newdata contains the peptide's isotope configuration.}

\item{dataCols}{\code{character} or \code{numeric} vector indicating which
columns of newdata contain the quantitative data. It should have the same
length as the number of samples in x (\code{ncol(x)}). Samples should be in
the same order.}

\item{idColPept}{\code{character} indicating which column contains the ids
that will be used in the merge with rowDataPept(x). The column in newdata
should have the same name as in rowDataPept(x).}
}
\value{
A \code{SilacPeptideExperiment} or \code{SilacProteomicsExperiment}
with new assay entries. If x is a \code{SilacProteinExperiment} then a
\code{SilacPeptideExperiment} is returned.
}
\description{
To estimate how much of the "old" isotope is being used in
"new" proteins the expression level of miss-cleaved peptides that contain
a mix of isotopes (one old and one new) and miss-cleaved peptides that
contain only new isotopes can be used to estimate amino acid recycling.
To add this information  a \code{data.frame} with the following information
is required:
\itemize{
  \item A column with ids that can be mapped to the peptide rowData (not
  necessary for \code{SilacProteinExperiment} objects).
  \item A column indicating which isotope configuration the peptide has: two
  heavy isotopes, mix of light and heavy isotope, etc.
  \item A set of columns with peptide expression quantification. The number
  of columns should be the same, and in the same order, as the number of
  samples in the \code{SilacProteinExperiment}, \code{SilacPeptideExperiment}
  or \code{SilacProteomicsExperiment}.
}
Non-detected measurements should be \code{NA}.
}
\examples{
data('wormsPE')
data('recycleLightLysine')
protPE <- ProtExp(wormsPE)
missPE <- addMisscleavedPeptides(x = protPE,
                                 newdata = recycleLightLysine,
                                 idColPept = 'Sequence',
                                 modCol = 'Modifications',
                                 dataCols = c(18:31))

names(assays(missPE))[1:2] <- c('int_lys8lys8', 'int_lys8lys0')
missPE <- calculateOldIsotopePool(x = missPE, 'int_lys8lys8', 'int_lys8lys0')

plotDistributionAssay(missPE, assayName = 'oldIsotopePool')

}
