% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotCounts.R
\docType{methods}
\name{barplotCounts}
\alias{barplotCounts}
\alias{barplotCounts,ProteinExperiment-method}
\alias{barplotCounts,PeptideExperiment-method}
\alias{barplotCounts,ProteomicsExperiment-method}
\title{Number of detected features per sample}
\usage{
barplotCounts(x, ...)

\S4method{barplotCounts}{ProteinExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)

\S4method{barplotCounts}{PeptideExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)

\S4method{barplotCounts}{ProteomicsExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{...}{Unused.}

\item{assayName}{Name of the assay to use in the plot.}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}
}
\value{
A ggplot2 barplot object or a \code{data.frame}.
}
\description{
How many proteins/peptides are detected in each sample.
\code{NA} are considered missing values.
}
\examples{
data('wormsPE')
barplotCounts(wormsPE, assayName = 'ratio')

}
