% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotTimeCoverage.R
\docType{methods}
\name{barplotTimeCoverage}
\alias{barplotTimeCoverage}
\alias{barplotTimeCoverage,SilacProteinExperiment-method}
\alias{barplotTimeCoverage,SilacPeptideExperiment-method}
\alias{barplotTimeCoverage,SilacProteomicsExperiment-method}
\title{Number of detected features per sample}
\usage{
barplotTimeCoverage(x, ...)

\S4method{barplotTimeCoverage}{SilacProteinExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)

\S4method{barplotTimeCoverage}{SilacPeptideExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)

\S4method{barplotTimeCoverage}{SilacProteomicsExperiment}(x, assayName,
  returnDataFrame = FALSE, conditionCol)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object.}

\item{...}{Unused.}

\item{assayName}{Name of the assay to use in the plot.}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}
}
\value{
A barplot or a \code{data.frame}.
}
\description{
How many proteins/peptides are detected in each sample. Anything
else than NA is considered detected.
}
\examples{
data('wormsPE')
barplotTimeCoverage(wormsPE, assayName = 'ratio')

}
