% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIsotopeFraction.R
\docType{methods}
\name{calculateIsotopeFraction}
\alias{calculateIsotopeFraction}
\alias{calculateIsotopeFraction,ProteinExperiment-method}
\alias{calculateIsotopeFraction,PeptideExperiment-method}
\alias{calculateIsotopeFraction,ProteomicsExperiment-method}
\title{Calculates the incorporated isotope fraction}
\usage{
calculateIsotopeFraction(x, ...)

\S4method{calculateIsotopeFraction}{ProteinExperiment}(x,
  ratioAssay = "ratio", oldIsoAssay, newIsoAssay, earlyTimepoints,
  lateTimepoints, conditionCol)

\S4method{calculateIsotopeFraction}{PeptideExperiment}(x,
  ratioAssay = "ratio", oldIsoAssay, newIsoAssay, earlyTimepoints,
  lateTimepoints, conditionCol)

\S4method{calculateIsotopeFraction}{ProteomicsExperiment}(x,
  ratioAssay = "ratio", oldIsoAssay, newIsoAssay, earlyTimepoints,
  lateTimepoints, conditionCol)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or
\code{ProteomicsExperiment} object.}

\item{...}{Unused.}

\item{ratioAssay}{A \code{character} with the assay name that has the ratio
data.}

\item{oldIsoAssay}{A \code{character} with the assay name that has the new
isotope intensity data.}

\item{newIsoAssay}{A \code{character} with the assay name that has the old
isotope intensity data.}

\item{earlyTimepoints}{A \code{numeric} indicating which timepoints should
be considered early.}

\item{lateTimepoints}{A \code{numeric} indicating which timepoints should
be considered late.}

\item{conditionCol}{A \code{character} indicating which column of the colData
data.frame indicates the different conditions.}
}
\value{
a \code{ProteinExperiment}, \code{PeptideExperiment} or
\code{ProteomicsExperiment} object with additional assays named "fraction".
}
\description{
Calculates the fraction of an isotope ratio using the following
formula:
\deqn{Isotope fraction_{A} = \frac{ratio}{ratio + 1}}
The ratio should be calculated as:
\deqn{ratio = isotope_{A}/isotope_{B}}
}
\details{
If oldIsoAssay and newIsoAssay arguments are given, then the
earlyTimepoints and lateTiempoints arguments can be used. These can be used
for example if certain proteins do not have any new isotope intensity during
the early timepoint. Because of that, no ratio can be calculated and could
lead to additional missing values. If old isotope intensity is detected, then
a fraction of 0 for new isotope is given. Same principle applies for the
late timepoint but with the isotopes in reverse.
}
\examples{
data('wormsPE')
calculateIsotopeFraction(wormsPE)

}
\seealso{
\code{\link{ProteomicsExperiment-class}}
}
