% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIsotopeRatio.R
\docType{methods}
\name{calculateIsotopeRatio}
\alias{calculateIsotopeRatio}
\alias{calculateIsotopeRatio,ProteinExperiment-method}
\alias{calculateIsotopeRatio,PeptideExperiment-method}
\alias{calculateIsotopeRatio,ProteomicsExperiment-method}
\title{Ratio calculation}
\usage{
calculateIsotopeRatio(x, newIsotopeAssay, oldIsotopeAssay, ...)

\S4method{calculateIsotopeRatio}{ProteinExperiment}(x, newIsotopeAssay,
  oldIsotopeAssay)

\S4method{calculateIsotopeRatio}{PeptideExperiment}(x, newIsotopeAssay,
  oldIsotopeAssay)

\S4method{calculateIsotopeRatio}{ProteomicsExperiment}(x, newIsotopeAssay,
  oldIsotopeAssay)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{newIsotopeAssay}{A \code{character} indicating the assay name that
has the new isotope intensity data.}

\item{oldIsotopeAssay}{A \code{character} indicating the assay name that
has the old isotope intensity data.}

\item{...}{Unused.}
}
\value{
A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object with an added assay named 'ratio'.
}
\description{
Calculates the ratio between the new isotope and the old
isotope (new/old).
}
\examples{
data('wormsPE')
calculateIsotopeRatio(x = wormsPE,
                      newIsotopeAssay = 'int_heavy',
                      oldIsotopeAssay = 'int_light')

}
