% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOldIsotopePool.R
\name{calculateOldIsotopePool}
\alias{calculateOldIsotopePool}
\alias{calculateOldIsotopePool,SilacPeptideExperiment-method}
\alias{calculateOldIsotopePool,SilacProteomicsExperiment-method}
\title{Estimates the fraction of old isotope for each time point}
\usage{
calculateOldIsotopePool(x, ...)

\S4method{calculateOldIsotopePool}{SilacPeptideExperiment}(x, newIsotopeAssayName, mixIsotopeAssayName)

\S4method{calculateOldIsotopePool}{SilacProteomicsExperiment}(x, newIsotopeAssayName, mixIsotopeAssayName)
}
\arguments{
\item{x}{A \code{SilacPeptideExperiment} or \code{SilacProteomicsExperiment}
object.}

\item{...}{Unused.}

\item{newIsotopeAssayName}{\code{character} indicating the assay that
contains quantification data for miss-cleaved peptides with two new isotopes
incorporated.}

\item{mixIsotopeAssayName}{\code{character} indicating the assay that
contains quantification data for miss-cleaved peptides with one old isotope
and one new isotope incorporated.}
}
\value{
A \code{SilacPeptideExperiment} or \code{SilacProteomicsExperiment}
with a peptide assay entry named "oldIsotopePool".
}
\description{
To estimate how much of the "old" isotope is being used in "new"
proteins we can use the expression level of miss-cleaved peptides that
contain a mix of isotopes (one old and one new) and miss-cleaved peptides
that contain only new isotopes. This can be done using the following formula:

\ifelse{html}{\out{<img src="https://latex.codecogs.com/svg.latex?\inline&space;\frac{1}{2\frac{Intensity_{lys8lys8}}{Intensity_{lys8lys0}}&space;&plus;&space;1}&space;=&space;lys_0&space;(Fraction)" title="\frac{1}{2\frac{Intensity_{lys8lys8}}{Intensity_{lys8lys0}} + 1} = lys_0 (Fraction)" />}}{\deqn{\frac{1}{2\frac{Intensity_{lys8lys8}}{Intensity_{lys8lys0}} + 1} = lys_0 (Fraction)}}

Which gives an idea of how much recyling (turnover understimation) is
happening.

Both peptide types, mix of old/new isotope and two new isotopes, have to be
found in a time point to calculate the fraction of old isotope.
}
\examples{
data('wormsPE')
data('recycleLightLysine')
protPE <- ProtExp(wormsPE)
missPE <- addMisscleavedPeptides(x = protPE,
                                 newdata = recycleLightLysine,
                                 idColPept = 'Sequence',
                                 modCol = 'Modifications',
                                 dataCols = c(18:31))

names(assays(missPE))[1:2] <- c('int_lys8lys8', 'int_lys8lys0')
missPE <- calculateOldIsotopePool(x = missPE, 'int_lys8lys8', 'int_lys8lys0')

plotDistributionAssay(missPE, assayName = 'oldIsotopePool')

}
