% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareAIC.R
\name{compareAIC}
\alias{compareAIC}
\title{Calculates the probability of each model from a set of models.}
\usage{
compareAIC(...)
}
\arguments{
\item{...}{a \code{list} with the model metrics,
the output from \code{\link{modelTurnover}} and \code{\link{calculateAIC}}.}
}
\value{
a \code{list} with a matrix for each experiment condition. The matrix
contains the probabilities of each model (columns) for each protein/peptide
(rows).
}
\description{
For a given set of AIC from models, the probability of each model relative
to the rest of the models of the set is calculated using the following
formula:
\deqn{\prod AIC_{i} = \frac{exp(\frac{AIC_{min}-AIC_{i}}{2})}{\sum_{j}exp(\frac{AIC_{min}-AIC_{j}}{2})}}
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList1 <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

modelList1 <- calculateAIC(modelList1, smallSampleSize = TRUE)

modelList2 <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t) + b',
                           start = list(k = 0.02, b = 0),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

modelList2 <- calculateAIC(modelList2, smallSampleSize = TRUE)

modelProbabilities <- compareAIC(modelList1, modelList2)

}
\seealso{
\code{\link{calculateAIC}},
         \code{\link{modelTurnover}}.
}
