% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterByMissingTimepoints.R
\docType{methods}
\name{filterByMissingTimepoints}
\alias{filterByMissingTimepoints}
\alias{filterByMissingTimepoints,SilacProteinExperiment-method}
\alias{filterByMissingTimepoints,SilacPeptideExperiment-method}
\alias{filterByMissingTimepoints,SilacProteomicsExperiment-method}
\title{Filter proteins/peptides by the amount of measurements overtime}
\usage{
filterByMissingTimepoints(x, ...)

\S4method{filterByMissingTimepoints}{SilacProteinExperiment}(x, assayName,
  maxMissing = 0, strict = TRUE, conditionCol, returnVector = FALSE)

\S4method{filterByMissingTimepoints}{SilacPeptideExperiment}(x, assayName,
  maxMissing = 0, strict = TRUE, conditionCol, returnVector = FALSE)

\S4method{filterByMissingTimepoints}{SilacProteomicsExperiment}(x,
  assayName, maxMissing = 0, strict = TRUE, conditionCol,
  returnVector = FALSE)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object.}

\item{...}{Unused.}

\item{assayName}{A character indicating which assay will be used to count
the number of missed timepoints.}

\item{maxMissing}{A numeric indicating how many timepoints are allowed to be
missed.}

\item{strict}{Logical: if TRUE, then proteins have to meet the maxMissing
criteria in all conditions and time replicates to pass; if FALSE then
proteins only have to meet the maxMissing criteria in one condition or time
replicate to pass.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{returnVector}{Logical: if TRUE then a vector with the positions to be
subset is returned.}
}
\value{
A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object or a logical vector with the rows that
pass the minimum number of desired timepoints.
}
\description{
Searches for proteins/peptides that are not found in all
timepoints. This can be done for each condition independently
(strict = FALSE) or shared across conditions (strict = TRUE).
}
\examples{
data('wormsPE')
filterByMissingTimepoints(wormsPE,
                          assayName = 'ratio',
                          maxMissing = 2,
                          strict = FALSE)
}
