% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mefPE}
\alias{mefPE}
\title{SilacProteomicsExperiment with pulsed silac data from MEFs}
\format{A \code{SilacProteinExperiment} object with 5223 proteins in a total
of 10 samples.
\describe{
  \item{colData}{A \code{DataFrame} with the design of the experiment:
  samples, timepoints, replicates...}
  \item{assays}{A list of matrices with quantification data at protein
  level: ratio and fraction.}
  \item{rowData}{A \code{DataFrame} with 3 columns that general protein
  id information.}
}}
\usage{
mefPE
}
\description{
A pre-built \code{SilacProteinExperiment} object with data from a pulsed
silac experiment done in mouse embryonic fibroblasts (MEFs). Two cell
cultures are compared: cultured with or without serum.
}
\details{
This dataset is used as an example, in the pulsed silac vignette, to show
the effect of comparing protein turnover between cell lines growing at
different rates.
}
\keyword{datasets}
