% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\docType{methods}
\name{merge}
\alias{merge}
\alias{merge,ProteinExperiment,ANY-method}
\alias{merge,PeptideExperiment,ANY-method}
\alias{merge,ProteomicsExperiment,ANY-method}
\title{Merge}
\usage{
\S4method{merge}{ProteinExperiment,ANY}(x, y, by, by.x = by, by.y = by,
  all = TRUE, ...)

\S4method{merge}{PeptideExperiment,ANY}(x, y, by, by.x = by, by.y = by,
  all = TRUE, ...)

\S4method{merge}{ProteomicsExperiment,ANY}(x, y, by.prot,
  by.prot.x = by.prot, by.prot.y = by.prot, by.pept,
  by.pept.x = by.pept, by.pept.y = by.pept, all = TRUE, ...)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{y}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{by, by.x, by.y}{A \code{character} indicating the columns used for the
merging.}

\item{all}{A \code{logical} indicating if all proteins/peptides should
be returned or only the intersect.}

\item{...}{Further parameters passed into \code{base::merge}.}

\item{by.prot, by.prot.x, by.prot.y}{For \code{ProteomicsExperiment} objects
a \code{character} indicating the columns used for the merging of the
protein level.}

\item{by.pept, by.pept.x, by.pept.y}{For \code{ProteomicsExperiment} objects
a \code{character} indicating the columns used for the merging of the
protein level.}
}
\value{
A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.
}
\description{
Merges two objects of the same class: \code{ProteinExperiment},
\code{PeptideExperiment} or \code{ProteomicsExperiment}.
}
\details{
This function is designed to be able to merge different samples
from different experiments since it is probable that not the exact same
proteins are found in both experiments and therefore \code{cbind} cannot be
used. It uses the merge base function to merge the rowData data frames and
merges the assays based on such merge. The colData \code{data.frame} are
joined.

For a \code{ProteomicsExperiment} object it gets a bit more complicated since
it is possible that some peptides that were assigned to one protein in
one experiment are assigned to another one in another experiment. Therefore
the linkerDf \code{data.frame} is recalculated.
}
\examples{
data('wormsPE')
merge(wormsPE[1:10, 1:3], wormsPE[3:10, 4:5])
}
