% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeModelsLists.R
\name{mergeModelsLists}
\alias{mergeModelsLists}
\title{Merge several models lists.}
\usage{
mergeModelsLists(...)
}
\arguments{
\item{...}{Lists with model data, output from \code{\link{modelTurnover}}.}
}
\value{
A list of models.
}
\description{
Merges several models lists into one. The models lists must come
from \code{\link{modelTurnover}} with the same arguments with the exception
of the input data. This function should be used in cases where a model is
fit for different conditions of an experiment with NO overlap in the samples
and NO missing samples. Otherwise the plotting functions might give
incorrect outputs.
}
\details{
When merging the attributes are also merged. Some of these need
to be recalculated since they contain information about the input data
positioning (column number). These attributes are used in the plotting
functions.

Take this into consideration so that the order of the models lists follows
the order of the columns in the original data. This also means no skipped
conditions and no skipped samples. If that is the case, build an
intermediari object that contains only the samples to be used (see examples).
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList1 <- modelTurnover(x = wormsPE[1:10, 1:7],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

modelList2 <- modelTurnover(x = wormsPE[1:10, 8:14],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

mergedModelList <- mergeModelsLists(modelList1, modelList2)

}
\seealso{
\code{\link{modelTurnover}}
}
