% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTurnover.R
\docType{methods}
\name{modelTurnover}
\alias{modelTurnover}
\alias{modelTurnover,ProteinExperiment-method}
\alias{modelTurnover,PeptideExperiment-method}
\alias{modelTurnover,ProteomicsExperiment-method}
\title{Estimate protein/peptide turnover}
\usage{
modelTurnover(x, ...)

\S4method{modelTurnover}{ProteinExperiment}(x, assayName = "fraction",
  formula = "fraction ~ 1-exp(-k*t)", start = list(k = 0.02),
  robust = FALSE, mode = "protein", verbose = FALSE,
  returnModel = FALSE, conditionCol, timeCol, ...)

\S4method{modelTurnover}{PeptideExperiment}(x, assayName = "fraction",
  formula = "fraction ~ 1-exp(-k*t)", start = list(k = 0.02),
  robust = FALSE, mode = c("grouped", "peptide"), verbose = FALSE,
  returnModel = FALSE, conditionCol, timeCol, proteinCol, ...)

\S4method{modelTurnover}{ProteomicsExperiment}(x, assayName = "fraction",
  formula = "fraction ~ 1-exp(-k*t)", start = list(k = 0.02),
  robust = FALSE, mode = c("protein", "grouped", "peptide"),
  verbose = FALSE, returnModel = FALSE, conditionCol, timeCol,
  proteinCol, ...)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or
\code{ProteomicsExperiment} object.}

\item{...}{further parameters passed into \code{nls} or \code{nlrob}.}

\item{assayName}{\code{character} indicating which assay to use as data
input for the model.}

\item{formula}{\code{formula} to be used. Time must always be named "t" and
the data must be named "fraction".}

\item{start}{named \code{list} with the initical values for the parameters
in formula.}

\item{robust}{\code{logical} indicating if robust modelling from the
\code{robustbase} package should be used.}

\item{mode}{\code{character} indicating which type of data should be used.
Can be "protein": one model per protein; "grouped": one model per protein
using peptide data; "peptide" one model per peptide.}

\item{verbose}{\code{logical} indicating if a progress bar should be
printed.}

\item{returnModel}{\code{logical} indicating if the model objects should
be returned also in the output.}

\item{conditionCol}{\code{character} indicating which column of colData(x)
describes the conditions.}

\item{timeCol}{\code{character} indicating which column of colData(x)
describes time.}

\item{proteinCol}{\code{character} indicating which column of rowData(x)
describes the assigned protein to a peptide. (Only for peptide data)}
}
\value{
A named \code{list} with either model metrics in matrices or the
model objects.
}
\description{
Method to apply turnover models on protein/peptide data
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

}
