% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptideexperiment.R
\name{PeptideExperiment}
\alias{PeptideExperiment}
\title{PeptideExperiment constructor}
\usage{
PeptideExperiment(assays, rowData = NULL, colData = NULL,
  conditionCol = NA, timeCol = NA, replicateIntCol = NA,
  replicateTimeCol = NA, metadata = NULL)
}
\description{
PeptideExperiment constructor
}
