% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributionModel.R
\name{plotDistributionModel}
\alias{plotDistributionModel}
\title{Distribution of modelling output}
\usage{
plotDistributionModel(
  modelList,
  value = "param_values",
  plotType = "density",
  returnDataFrame = FALSE
)
}
\arguments{
\item{modelList}{A list containing all the model objects, this should be the
output of \code{\link{modelTurnover}}.}

\item{value}{A \code{character} indicating which metric to plot.  Check
\code{names(modelList)} for available options. (Default = 'param_values')}

\item{plotType}{A \code{character} indicating which geometry to plot:
'boxplot' or 'density'. (default = 'density')}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}
}
\value{
A \code{ggplot} density or boxplot object, or the \code{data.frame}
used to make the plot.
}
\description{
Plot the distribution of the different model parameters and
metrics for each condition.
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

plotDistributionModel(modelList = modelList,
                      value = 'param_values',
                      plotType = 'density')

}
