% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualModel.R
\docType{methods}
\name{plotIndividualModel}
\alias{plotIndividualModel}
\alias{plotIndividualModel,SilacProteinExperiment-method}
\alias{plotIndividualModel,SilacPeptideExperiment-method}
\alias{plotIndividualModel,SilacProteomicsExperiment-method}
\title{Fitted model(s) for a feature}
\usage{
plotIndividualModel(x, ...)

\S4method{plotIndividualModel}{SilacProteinExperiment}(x, modelList, num,
  returnDataFrame = FALSE)

\S4method{plotIndividualModel}{SilacPeptideExperiment}(x, modelList, num,
  returnDataFrame = FALSE)

\S4method{plotIndividualModel}{SilacProteomicsExperiment}(x, modelList,
  num, returnDataFrame = FALSE)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or
\code{SilacProteomicsExperiment} object.}

\item{...}{Unused.}

\item{modelList}{A list containing all the model objects, this should be the
output of \code{\link{modelTurnover}} with returnModel as TRUE.}

\item{num}{The feature number to be plotted.}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}
}
\value{
A scatter plot with a fitted line or a \code{data.frame}.
}
\description{
Plot the model fit for a specific protein/peptide in different
conditions.
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

plotIndividualModel(x = wormsPE,
                    modelList = modelList,
                    num = 2)

}
