% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ProteinExperiment-class}
\alias{ProteinExperiment-class}
\alias{.ProteinExperiment}
\title{ProteinExperiment class}
\description{
S4 class that extends the \code{\link{SummarizedExperiment}}
class. This class is designed for proteomics data, more especifically
protein level data. The \code{metadata} slot comes already initialized with
the metaoptions (see details).
}
\details{
The \code{ProteinExperiment} class has been designed to store
protein level data and to be used in the functions provided in this package
for pulsed SILAC data analysis; in combination with the other two classes
from the package: the \code{\link{PeptideExperiment}} and
\code{\link{ProteomicsExperiment}} classes.

ProteinExperiment metaoptions are stored in the \code{metadata} slot
This contains a \code{list} with some parameters that are automatically
initialized by the constructor. Some parameters are mandatory for certain
functions or operations. The user can add or remove items at their
discretion. These parameters are meant to help automate certain pipeline or
data analysis steps. These metaoptions are:
\describe{
  \item{conditionCol}{\code{character} indicating the column name of
  \code{colData(x)} that defines the different experiment conditions.}
  \item{timeCol}{\code{character} indicating the column name of
  \code{colData(x)} that defines the different timepoints of the experiment.}
}
}
\section{Constructor}{

See \link{ProteinExperiment-constructor} for details.
}

\section{Accessors}{

See \link{ProteinPeptideExperiment-accessors} for details.
}

\seealso{
\code{\link{ProteinExperiment-constructor}},
         \code{\link{ProteinPeptideExperiment-accessors}},
         \code{\link{SummarizedExperiment}}
}
