% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinExperiment.R
\name{ProteinExperiment-constructor}
\alias{ProteinExperiment-constructor}
\alias{ProteinExperiment}
\title{ProteinExperiment constructor}
\usage{
ProteinExperiment(assays, rowData = NULL, colData = NULL,
  conditionCol = NA, timeCol = NA, metadata = NULL,
  metaoptions = NULL)
}
\arguments{
\item{assays}{A named \code{list} of matrices (assays) with protein level
data.}

\item{rowData}{A \code{data.frame} with protein feature data
like protein names, molecular weight, etc.}

\item{colData}{A \code{data.frame} with sample information like
conditions, replicates, etc.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{timeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment timepoints.}

\item{metadata}{A \code{list} to store any kind of experiment-wide
data; like authors, dates, machines used...}

\item{metaoptions}{A \code{list} to store user defined metaoptions.}
}
\value{
An object of class \code{ProteinExperiment}.
}
\description{
Constructor function for the ProteinExperiment class object.
}
\section{Class description}{

See \link{ProteinExperiment-class} for details.
}

\section{Accessors}{

See \link{ProteinPeptideExperiment-accessors} for details.
}

\examples{

## assays
assays_protein <- list(expression = matrix(1:9, ncol = 3))

## colData
colData <- data.frame(sample = c('A1', 'A2', 'A3'),
                      condition = c('A', 'A', 'A'),
                    time = c(1, 2, 3))
## rowData
rowData_protein <- data.frame(prot_id = LETTERS[1:3])

## construct the ProteinExperiment
protExp <- ProteinExperiment(assays = assays_protein,
                             rowData = rowData_protein,
                             colData = colData,
                             conditionCol = 'condition',
                             timeCol = 'time')

}
