% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinexperiment.R
\name{ProteinExperiment}
\alias{ProteinExperiment}
\title{ProteinExperiment constructor}
\usage{
ProteinExperiment(assays, rowData = NULL, colData = NULL,
  conditionCol = NA, timeCol = NA, replicateIntCol = NA,
  replicateTimeCol = NA, metadata = NULL)
}
\description{
ProteinExperiment constructor
}
