% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assays.R, R/cbind.R, R/coercers.R, R/coldata.R,
%   R/metaoptions.R, R/rbind.R, R/rowdata.R, R/subset.R, R/zzz.R
\docType{methods}
\name{ProteinPeptideExperiment-accessors}
\alias{ProteinPeptideExperiment-accessors}
\alias{assays<-,ProteinExperiment,ANY-method}
\alias{assays<-,PeptideExperiment,ANY-method}
\alias{cbind,ProteinExperiment-method}
\alias{cbind,PeptideExperiment-method}
\alias{coerce}
\alias{coerce,ProteinExperiment,SummarizedExperiment-method}
\alias{coerce,ProteinExperiment,data.frame-method}
\alias{colData<-,ProteinExperiment,ANY-method}
\alias{colData<-,ProteinExperiment-method}
\alias{colData<-,PeptideExperiment,ANY-method}
\alias{colData<-,PeptideExperiment-method}
\alias{metaoptions,ProteinExperiment-method}
\alias{metaoptions<-,ProteinExperiment-method}
\alias{metaoptions,PeptideExperiment-method}
\alias{metaoptions<-,PeptideExperiment-method}
\alias{rbind,ProteinExperiment-method}
\alias{rbind,PeptideExperiment-method}
\alias{rowData<-,ProteinExperiment-method}
\alias{rowData<-,PeptideExperiment-method}
\alias{subset,ProteinExperiment-method}
\alias{subset,PeptideExperiment-method}
\title{Accessors for the ProteinExperiment and PeptideExperiment classes}
\usage{
\S4method{assays}{ProteinExperiment,ANY}(x) <- value

\S4method{assays}{PeptideExperiment,ANY}(x) <- value

\S4method{cbind}{ProteinExperiment}(..., deparse.level = 1)

\S4method{cbind}{PeptideExperiment}(..., deparse.level = 1)

\S4method{colData}{ProteinExperiment,ANY}(x, ...) <- value

\S4method{colData}{PeptideExperiment,ANY}(x, ...) <- value

\S4method{metaoptions}{ProteinExperiment}(x)

\S4method{metaoptions}{ProteinExperiment}(x) <- value

\S4method{metaoptions}{PeptideExperiment}(x)

\S4method{metaoptions}{PeptideExperiment}(x) <- value

\S4method{rbind}{ProteinExperiment}(..., deparse.level = 1)

\S4method{rbind}{PeptideExperiment}(..., deparse.level = 1)

\S4method{rowData}{ProteinExperiment}(x, ...) <- value

\S4method{rowData}{PeptideExperiment}(x, ...) <- value

\S4method{subset}{ProteinExperiment}(x, ...)

\S4method{subset}{PeptideExperiment}(x, ...)
}
\arguments{
\item{x}{A \code{ProteinExperiment} or a \code{PeptideExperiment} object.}

\item{value}{An object of class specified in the S4 method signature or as
described in the following sections.}

\item{...}{For \code{rbind} and \code{cbind} are \code{ProteinExperiment} or
\code{PeptideExperiment} objects to be joined together. For \code{subset}
it is a logical comparison using a column name from the respective rowData
\code{data.frame}. Otherwise unused.}

\item{deparse.level}{Unused.}
}
\value{
Elements from \code{ProteinExperiment} or \code{PeptideExperiment}
objects.
}
\description{
All the accessors, dimension, subsetting, merging and coercers
that work on \code{ProteinExperiment} and \code{PeptideExperiment} objects.
Functions that work on \code{SummarizedExperiment} objects should also work
on these two objects. Detailed examples of these functions can be found in
the vignette of this package.
}
\section{Accessors}{


The following functions can be used to access the data in the class slots

\describe{
  \item{\code{assays}:}{Access the assays (list of matrices) of the object.
  Value should be a matrix or list of matrices.}
  \item{\code{assayNames}}{ Access the assay names of the object. Value
  should be a character vector.}
  \item{\code{rowData}}{Access the protein/peptide feature \code{data.frame}
  of the object. Value should be \code{data.frame} with as many rows as
  proteins/peptides.}
  \item{\code{colData}:}{Access the samples \code{data.frame} of the object.
  Value should be a \code{data.frame} with as many rows as samples.}
  \item{\code{metadata}:}{Access the metadata \code{list} of the object.
  Value should be a \code{list}.}
  \item{\code{metaoptions}:}{Access the metaoptions \code{list} of the
  object. Value should be a \code{list}.}
}
}

\section{Dimensions}{


The following functions can be used to get the number of proteins/peptides
and number of samples:

\describe{
  \item{\code{nrow}:}{Gives how many proteins and/or peptides the object
  has.}
  \item{\code{ncol}:}{Gives how many samples the object has.}
  \item{\code{dim}:}{Gives the number of proteins/peptides and the
  number of samples the object has.}
  \item{\code{length}:}{Gives how many proteins and/or peptides the object
  has.}
}
}

\section{Subsetting}{


The following functions can be used to subset the different classes:

\describe{
  \item{\code{$}:}{Gives a column from colData by name.}
  \item{\code{`[`}:}{Can be used to subset by row and column.}
  \item{\code{subset}:}{Allows to subset based on a logical comparison
  using a column name from the rowData \code{data.frame}.}
}
}

\section{Merging}{


The following functions can be used to aggregate objects of the same class
together:

\describe{
  \item{\code{cbind}:}{Joins two or more objects horizontally (adding
  samples). Must have the same proteins/peptides and in the same order.}
  \item{\code{rbind}:}{Joins two or more objects vertically (adding
  proteins/peptides). Must have the same samples and in the same order.}
  \item{\code{merge}:}{Joins two objects by adding new samples and tries
  to merge the proteins/peptide rowData data.frames.}
}

Merge methods are explained in detail in \link{merge}.
}

\section{Coercers}{


The folloing functions can be used to transform a \code{ProteinExperiment}
or a \code{PeptideExperiment} into a \code{SummarizedExperiment} or a
\code{data.frame}.

\describe{
  \item{\code{as(x, 'SummarizedEpriment')}:}{Transforms the object into an
  object of class \code{SummarizedExperiment}.}
  \item{\code{as(x, 'data.frame')}:}{Transforms the object into an
  object of class \code{data.frame}.}
}
}

\examples{
data('wormsPE')
protPE <- ProtExp(wormsPE)

# Accessors
## assays
assays(protPE)

## assaysNames
assayNames(protPE)

## colData
colData(protPE)

## rowData
rowData(protPE)

## metadata
metadata(protPE)

## metaoptions
#metaoptions(protPE)


# Dimensions
nrow(protPE)
ncol(protPE)
dim(protPE)
length(protPE)

# Subsetting
protPE$line
protPE[1,1]
subset(protPE, protein_id == 'AC3.2')

# Merging
rbind(protPE[1:10, ], protPE[11:20, ])
cbind(protPE[,1:2], protPE[,3:4])
#merge(protPE[1:10, 1:3], protPE[3:10, 4:5])

# Coercers
as(protPE, 'SummarizedExperiment')
as(protPE, 'data.frame')
}
