% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterCompareAssay.R
\docType{methods}
\name{scatterCompareAssays}
\alias{scatterCompareAssays}
\alias{scatterCompareAssays,SilacProteinExperiment-method}
\alias{scatterCompareAssays,SilacPeptideExperiment-method}
\alias{scatterCompareAssays,SilacProteomicsExperiment-method}
\title{Scatter plot of two conditions for each timepoint of an assay.}
\usage{
scatterCompareAssays(x, ...)

\S4method{scatterCompareAssays}{SilacProteinExperiment}(x, conditions,
  assayName, returnDataFrame = FALSE, conditionCol, timeCol)

\S4method{scatterCompareAssays}{SilacPeptideExperiment}(x, conditions,
  assayName, returnDataFrame = FALSE, conditionCol, timeCol)

\S4method{scatterCompareAssays}{SilacProteomicsExperiment}(x, conditions,
  assayName, mode = "protein", returnDataFrame = FALSE, conditionCol,
  timeCol)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object.}

\item{...}{Unused.}

\item{conditions}{A \code{character} of length 2 indicating which 2
conditions should be compared.}

\item{assayName}{Name of the assay to use in the plot.}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{timeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different timepoints.}

\item{mode}{A \code{character} indicating which level of data to use,
either "protein" or "peptide". Only relevant for ProteomicsExperiment
inputs.}
}
\value{
A \code{ggplot} object or the \code{data.frame} that would be used
instead in the plot.
}
\description{
Scatter plot of two conditions/replicates for a selected assay.
Timepoints are separated using \code{facet_wrap}.
}
\examples{
data('wormsPE')
scatterCompareAssays(x = wormsPE,
                     conditions = c('OW40', 'OW450'),
                     assayName = 'ratio',
                     mode = 'protein')

}
