% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterCompareModels.R
\name{scatterCompareModels}
\alias{scatterCompareModels}
\title{Scatter plot of two conditions for a model metric.}
\usage{
scatterCompareModels(modelList, conditions, value = "param_values",
  returnDataFrame = FALSE)
}
\arguments{
\item{modelList}{A list containing all the model objects, this should be the
output of \code{\link{modelTurnover}} with returnModel as TRUE.}

\item{conditions}{A \code{character} of length 2 indicating which 2
conditions should be compared.}

\item{value}{A \code{character} indicating which metric to plot.  Check
\code{names(modelList)} for available options. (Default = 'param_values')}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}
}
\value{
A \code{ggplot} object or the \code{data.frame} that would be used
instead in the plot.
}
\description{
Scatter plot of two conditions/replicates for a selected metric
of a model. For example to compare turnover rates, model errors...
\code{facet_wrap}.
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

scatterCompareModels(modelList = modelList,
                     conditions = c('OW40', 'OW450'),
                     value = 'param_values')

}
