% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideExperiment.R, R/ProteinExperiment.R,
%   R/assayNames.R, R/assays.R, R/cbind.R, R/coldata.R, R/colnames.R, R/dim.R,
%   R/dolar.R, R/length.R, R/linkerDf.R, R/metadata.R, R/metaoptions.R,
%   R/ncol.R, R/nrow.R, R/rbind.R, R/rowdata.R, R/rownames.R, R/subset.R,
%   R/zzz.R
\name{SilacProteomicsExperiment-accessors}
\alias{SilacProteomicsExperiment-accessors}
\alias{PeptExp}
\alias{ProtExp}
\alias{assayNames,SilacProteomicsExperiment-method}
\alias{assayNamesProt}
\alias{assayNamesProt,SilacProteinExperiment-method}
\alias{assayNamesProt<-}
\alias{assayNamesProt<-,SilacProteinExperiment-method}
\alias{assayNamesProt,SilacProteomicsExperiment-method}
\alias{assayNamesProt<-,SilacProteomicsExperiment-method}
\alias{assayNamesPept}
\alias{assayNamesPept,SilacPeptideExperiment-method}
\alias{assayNamesPept<-}
\alias{assayNamesPept<-,SilacPeptideExperiment-method}
\alias{assayNamesPept,SilacProteomicsExperiment-method}
\alias{assayNamesPept<-,SilacProteomicsExperiment-method}
\alias{assays}
\alias{assays,SilacProteomicsExperiment-method}
\alias{assaysProt}
\alias{assaysProt,SilacProteinExperiment-method}
\alias{assaysProt<-}
\alias{assaysProt<-,SilacProteinExperiment-method}
\alias{assaysProt,SilacProteomicsExperiment-method}
\alias{assaysProt<-,SilacProteomicsExperiment-method}
\alias{assaysPept}
\alias{assaysPept,SilacPeptideExperiment-method}
\alias{assaysPept<-}
\alias{assaysPept<-,SilacPeptideExperiment-method}
\alias{assaysPept,SilacProteomicsExperiment-method}
\alias{assaysPept<-,SilacProteomicsExperiment-method}
\alias{cbind,SilacProteomicsExperiment-method}
\alias{colData,SilacProteomicsExperiment-method}
\alias{colData<-,SilacProteomicsExperiment,ANY-method}
\alias{colData<-,SilacProteomicsExperiment-method}
\alias{colnames,SilacProteomicsExperiment-method}
\alias{colnames<-,SilacProteomicsExperiment-method}
\alias{dim,SilacProteomicsExperiment-method}
\alias{$,SilacProteomicsExperiment-method}
\alias{$<-,SilacProteomicsExperiment-method}
\alias{length,SilacProteomicsExperiment-method}
\alias{linkerDf}
\alias{linkerDf,SilacProteomicsExperiment-method}
\alias{linkerDf<-}
\alias{linkerDf<-,SilacProteomicsExperiment-method}
\alias{metadata,SilacProteomicsExperiment-method}
\alias{metadata<-,SilacProteomicsExperiment-method}
\alias{metaoptions}
\alias{metaoptions,SilacProteomicsExperiment-method}
\alias{metaoptions<-}
\alias{metaoptions<-,SilacProteomicsExperiment-method}
\alias{ncol,SilacProteomicsExperiment-method}
\alias{ncol}
\alias{nrow,SilacProteomicsExperiment-method}
\alias{rbind,SilacProteomicsExperiment-method}
\alias{rowData,SilacProteomicsExperiment-method}
\alias{rowData<-,SilacProteomicsExperiment-method}
\alias{rowDataProt}
\alias{rowDataProt,SilacProteinExperiment-method}
\alias{rowDataProt<-}
\alias{rowDataProt<-,SilacProteinExperiment-method}
\alias{rowDataProt,SilacProteomicsExperiment-method}
\alias{rowDataProt<-,SilacProteomicsExperiment-method}
\alias{rowDataPept}
\alias{rowDataPept,SilacPeptideExperiment-method}
\alias{rowDataPept<-}
\alias{rowDataPept<-,SilacPeptideExperiment-method}
\alias{rowDataPept,SilacProteomicsExperiment-method}
\alias{rowDataPept<-,SilacProteomicsExperiment-method}
\alias{rownamesProt}
\alias{rownamesProt,SilacProteomicsExperiment-method}
\alias{rownamesProt<-}
\alias{rownamesProt<-,SilacProteomicsExperiment-method}
\alias{rownamesPept}
\alias{rownamesPept,SilacProteomicsExperiment-method}
\alias{rownamesPept<-}
\alias{rownamesPept<-,SilacProteomicsExperiment-method}
\alias{subset,SilacProteomicsExperiment-method}
\alias{subsetProt}
\alias{subsetProt,SilacProteinExperiment-method}
\alias{subsetProt,SilacProteomicsExperiment-method}
\alias{subsetPept}
\alias{subsetPept,SilacPeptideExperiment-method}
\alias{subsetPept,SilacProteomicsExperiment-method}
\alias{[,SilacProteomicsExperiment,ANY,ANY,ANY-method}
\alias{[,SilacProteomicsExperiment-ANY-ANY-method}
\title{Accessors for the SilacProteomicsExperiment class}
\usage{
PeptExp(x)

ProtExp(x)

\S4method{assayNames}{SilacProteomicsExperiment}(x, ..., withDimnames)

\S4method{assayNamesProt}{SilacProteinExperiment}(x)

\S4method{assayNamesProt}{SilacProteinExperiment}(x) <- value

\S4method{assayNamesProt}{SilacProteomicsExperiment}(x)

\S4method{assayNamesProt}{SilacProteomicsExperiment}(x) <- value

\S4method{assayNamesPept}{SilacPeptideExperiment}(x)

\S4method{assayNamesPept}{SilacPeptideExperiment}(x) <- value

\S4method{assayNamesPept}{SilacProteomicsExperiment}(x)

\S4method{assayNamesPept}{SilacProteomicsExperiment}(x) <- value

\S4method{assays}{SilacProteomicsExperiment}(x, ..., withDimnames = TRUE)

\S4method{assaysProt}{SilacProteinExperiment}(x)

\S4method{assaysProt}{SilacProteinExperiment}(x) <- value

\S4method{assaysProt}{SilacProteomicsExperiment}(x)

\S4method{assaysProt}{SilacProteomicsExperiment}(x) <- value

\S4method{assaysPept}{SilacPeptideExperiment}(x)

\S4method{assaysPept}{SilacPeptideExperiment}(x) <- value

\S4method{assaysPept}{SilacProteomicsExperiment}(x)

\S4method{assaysPept}{SilacProteomicsExperiment}(x) <- value

\S4method{cbind}{SilacProteomicsExperiment}(..., deparse.level = 1)

\S4method{colData}{SilacProteomicsExperiment}(x, ...)

\S4method{colData}{SilacProteomicsExperiment,ANY}(x, ...) <- value

\S4method{colnames}{SilacProteomicsExperiment}(x)

\S4method{colnames}{SilacProteomicsExperiment}(x) <- value

\S4method{dim}{SilacProteomicsExperiment}(x)

\S4method{$}{SilacProteomicsExperiment}(x, name)

\S4method{$}{SilacProteomicsExperiment}(x, name) <- value

\S4method{length}{SilacProteomicsExperiment}(x)

\S4method{linkerDf}{SilacProteomicsExperiment}(x)

\S4method{linkerDf}{SilacProteomicsExperiment}(x) <- value

\S4method{metadata}{SilacProteomicsExperiment}(x, ...)

\S4method{metadata}{SilacProteomicsExperiment}(x, ...) <- value

\S4method{metaoptions}{SilacProteomicsExperiment}(x)

\S4method{metaoptions}{SilacProteomicsExperiment}(x) <- value

\S4method{ncol}{SilacProteomicsExperiment}(x)

\S4method{rbind}{SilacProteomicsExperiment}(..., deparse.level = 1)

\S4method{rowData}{SilacProteomicsExperiment}(x, use.names = TRUE, ...)

\S4method{rowData}{SilacProteomicsExperiment}(x, ...) <- value

\S4method{rowDataProt}{SilacProteinExperiment}(x)

\S4method{rowDataProt}{SilacProteinExperiment}(x) <- value

\S4method{rowDataProt}{SilacProteomicsExperiment}(x)

\S4method{rowDataProt}{SilacProteomicsExperiment}(x) <- value

\S4method{rowDataPept}{SilacPeptideExperiment}(x)

\S4method{rowDataPept}{SilacPeptideExperiment}(x) <- value

\S4method{rowDataPept}{SilacProteomicsExperiment}(x)

\S4method{rownamesProt}{SilacProteomicsExperiment}(x)

\S4method{rownamesProt}{SilacProteomicsExperiment}(x) <- value

\S4method{rownamesPept}{SilacProteomicsExperiment}(x)

\S4method{rownamesPept}{SilacProteomicsExperiment}(x) <- value

\S4method{subset}{SilacProteomicsExperiment}(x, ...)

\S4method{subsetProt}{SilacProteinExperiment}(x, ...)

\S4method{subsetProt}{SilacProteomicsExperiment}(x, ...)

\S4method{subsetPept}{SilacPeptideExperiment}(x, ...)

\S4method{subsetPept}{SilacProteomicsExperiment}(x, ...)

\S4method{[}{SilacProteomicsExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A \code{SilacSilacProteomicsExperiment} object.}

\item{...}{For \code{rbind} and \code{cbind} are \code{SilacProteinExperiment} or
\code{SilacPeptideExperiment} objects to be joined together. For \code{subset}
it is a logical comparison using a column name from the respective rowData
\code{data.frame}. Otherwise unused.}

\item{withDimnames}{Unused.}

\item{value}{An object of class specified in the S4 method signature or as
described in the following sections.}

\item{deparse.level}{Unused.}

\item{name}{Column name of colData.}

\item{use.names}{Unused.}

\item{i, j}{For \code{`[`}, \code{i}, \code{j} are subscripts that can act to
subset the rows and columns of \code{x}.}

\item{drop}{A \code{logical} indicating if dimensions should be lowered if
possible when subsetting.}
}
\value{
Elements from a \code{SilacProteomicsExperiment} object.
}
\description{
All the accessors, dimension, subsetting, merging and coercers
that work on \code{SilacProteomicsExperiment} objects. Since the
\code{SilacProteomicsExperiment} object has both protein and peptide level data,
most of the functions have a 'Prot' or 'Pept' suffix to indicate which
level should be used. If the non-suffix function is used, then a list with
both protein and peptide data is returned.
These functions also work on \code{SilacProteinExperiment}
and \code{SilacPeptideExperiment} objects.
}
\section{Accessors}{


The following functions can be used to access the data in the class slots

\describe{
  \item{\code{assays}, \code{assaysProt}, \code{assaysPept}:}{Access the
   assays (list of matrices) of the object. Value should be a matrix or
   list of matrices.}
  \item{\code{assayNames}, \code{assayNamesProt}, \code{assayNamesPept}:}{
  Access the assay names of the object. Value should be a character vector.}
  \item{\code{rowData}, \code{rowDataProt}, \code{rowDataPept}:}{Access the
  protein/peptide feature \code{data.frame} of the object. Value should be a
  \code{data.frame} with as many rows as proteins/peptides.}
  \item{\code{colData}:}{Access the samples \code{data.frame} of the object.
  Value should be a \code{data.frame} with as many rows as samples.}
  \item{\code{metadata}:}{Access the metadata \code{list} of the object.
  Value should be a \code{list}.}
  \item{\code{metaoptions}:}{Access the metaoptions \code{list} of the
  object. Value should be a \code{list}.}
  \item{\code{linkerDf}:}{Access the linker \code{data.frame} of the object
  (only ProteomicsExperiment). Value should be a \code{data.frame} output
  from \code{\link{buildLinkerDf}}.}
  \item{\code{ProtExp} and \code{PeptExp}:}{Access the experiment objects
  in a \code{SilacProteomicsExperiment}.}
}
}

\section{Dimensions}{


The following functions can be used to get the number of proteins/peptides
and number of samples:

\describe{
  \item{\code{nrow}:}{Gives how many proteins and peptides the object
  has.}
  \item{\code{ncol}:}{Gives how many samples the object has.}
  \item{\code{dim}:}{Gives both the number of proteins andpeptides and the
  number of samples the object has.}
  \item{\code{length}:}{Gives how many proteins and or peptides the object
  has.}
}
}

\section{Subsetting}{


The following functions can be used to subset the different classes:

\describe{
  \item{\code{$}:}{Gives a column from colData by name.}
  \item{\code{`[`}:}{Can be used to subset by row and column.}
  \item{\code{subset}, \code{subsetProt} and \code{subsetPept}:}{Allows
  to subset based on a logical comparison using a column name from the
  rowData \code{data.frame}.}
}

The ProteomicsExperiment class is a bit more complex since there are two
levels at which the subset can be done and these two levels can be linked or
not.

If the metaoption 'linkedSubset' is TRUE, then when subsetting on one level,
the proteins/peptide linked to such level are also subsetted. Otherwise, one
of the levels remains unmodified.

\code{subsetProt} can be used to apply \code{subset} at the rowData
\code{data.frame} of the protein level. \code{subsetProt} can be used to
apply \code{subset} at the rowData \code{data.frame} of the peptide level. If
\code{subset} is used, then \code{subsetProt} or \code{subsetPept} will be
used depending on the metaoption 'subsetMode'.

`[` acts in the same manner as calling \code{subset}. In this case numerics
are used and samples can also be selected.

The vignette offers a detailed set of simple examples with all the
possible cases.
}

\section{Merging}{


The following functions can be used to aggregate objects of the same class
together:

\describe{
  \item{\code{cbind}:}{Joins two or more objects horizontally (adding
  samples). Must have the same proteins/peptides and in the same order.}
  \item{\code{rbind}:}{Joins two or more objects vertically (adding
  proteins/peptides). Must have the same samples and in the same order.}
  \item{\code{merge}:}{Joins two objects by adding new samples and tries
  to merge the proteins/peptide rowData data.frames and recalculate the
  linkerDf \code{data.frame} for the \code{SilacProteomicsExperiment} class.}
}

Merge methods are explained in detail in \link{merge}.
}

\examples{
# Accessors
## assays
data('wormsPE')
assays(wormsPE)
assaysProt(wormsPE)
assaysPept(wormsPE)

## assaysNames
assayNames(wormsPE)
assayNamesProt(wormsPE)
assayNamesPept(wormsPE)

## colData
colData(wormsPE)

## rowData
rowData(wormsPE)
rowDataProt(wormsPE)
rowDataPept(wormsPE)

## metadata
metadata(wormsPE)

## metaoptions
#metaoptions(wormsPE)

## linkerDf
linkerDf(wormsPE)

# Dimensions and dimensions names
nrow(wormsPE)
ncol(wormsPE)
dim(wormsPE)
length(wormsPE)
colnames(wormsPE)
rownamesProt(wormsPE)
rownamesPept(wormsPE)

# Subsetting
wormsPE$line
wormsPE[1,1]
subsetProt(wormsPE, protein_id == 'AC3.2')
subsetPept(wormsPE, Sequence == 'AIQEISDYHFLIK')

# Merging
rbind(wormsPE[1:10, ], wormsPE[11:20, ])
cbind(wormsPE[,1:2], wormsPE[,3:4])
#merge(wormsPE[1:10, 1:3], wormsPE[3:10, 4:5])
}
