% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SilacProteomicsExperiment-class}
\alias{SilacProteomicsExperiment-class}
\alias{.SilacProteomicsExperiment}
\title{SilacProteomicsExperiment class}
\description{
S4 class that contains a \code{SilacProteinExperiment} object
and a \code{SilacPeptideExperiment} object. The two objects are linked by a
\code{data.frame} (linkerDf). This class can be used to manage both protein
and peptide data at the same time.
}
\details{
The \code{SilacProteomicsExperiment} object is just a
\code{SilacProteinExperiment} object and a \code{SilacPeptideExperiment}
object together.

The rows of the \code{SilacProteinExperiment} object represent proteins.
The rows of the \code{SilacPeptideExperiment} object represent peptides.

The columns of the \code{SilacProteomicsExperiment} object represent samples.
Samples are shared at both protein and peptide levels.

Experiment-wide information can be stored in the \code{metadata} slot, which
is accessed with the \code{metadata} function. This contains a \code{list}
object in which each item is left to the discretion of the user. Some
possible examples could be: data of the experiment, author, machine used,
etc.

SilacProteomicsExperiment options are stored in the \code{metadata} slot.
This contains a \code{list} with some parameters that are automatically
initialized by the constructor. Some parameters are mandatory for certain
functions or operations. The user can add or remove items at their
discretion. These parameters are meant to help automate certain pipeline or
data analysis steps. These metaoptions are:
These metaoptions are:
\describe{
  \item{conditionCol}{\code{character} indicating the column name of
  \code{colData(x)} that defines the different experiment conditions.}
  \item{timeCol}{\code{character} indicating the column name of
  \code{colData(x)} that defines the different timepoints of the experiment.}
  \item{idColProt}{A \code{character} indicating which column from the
  rowData (protein) should be used as ids.}
  \item{idColPept}{A \code{character} indicating which column from the
  rowData (peptides) should be used as ids.}
  \item{linkedSubset}{A \code{logical} if subsetting should be linked between
  proteins and peptide.}
  \item{subsetMode}{A \code{character}, either 'protein' or 'peptide'
  indicating which level should be used first when subsetting.}
}
}
\section{Slots}{

\describe{
\item{\code{SilacProteinExperiment}}{Contains a \code{SilacProteinExperiment} object.}

\item{\code{SilacPeptideExperiment}}{Contains a \code{SilacPeptideExperiment} object.}

\item{\code{colData}}{Contains a \code{data.frame} with sample information like
conditions, replicates, etc.}

\item{\code{linkerDf}}{Contains a \code{data.frame} that has been created with
\code{\link{buildLinkerDf}}. It contains the relationships between proteins
and peptides.}

\item{\code{metadata}}{Contains a \code{list} to store any kind of experiment-wide
data and the metaoptions.}
}}

\section{Constructor}{

See \link{SilacProteomicsExperiment-constructor} for details.
}

\section{Accessors}{

See \link{SilacProteomicsExperiment-accessors} for details.
}

\seealso{
\code{\link{SilacProteomicsExperiment-constructor}},
         \code{\link{SilacProteomicsExperiment-accessors}},
         \code{\link{SilacProteinExperiment}},
         \code{\link{SilacPeptideExperiment}}
}
