% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteomicsExperiment.R
\name{SilacProteomicsExperiment-constructor}
\alias{SilacProteomicsExperiment-constructor}
\alias{SilacProteomicsExperiment}
\title{SilacProteomicsExperiment constructor}
\usage{
SilacProteomicsExperiment(
  SilacProteinExperiment,
  SilacPeptideExperiment,
  colData,
  linkerDf,
  metadata,
  idColProt = NA,
  idColPept = NA,
  linkedSubset = TRUE,
  subsetMode = "protein",
  conditionCol = NA,
  timeCol = NA,
  proteinCol = NA
)
}
\arguments{
\item{SilacProteinExperiment}{A \code{SilacProteinExperiment} object.}

\item{SilacPeptideExperiment}{A \code{SilacPeptideExperiment} object.}

\item{colData}{A \code{data.frame} with sample information like
conditions, replicates, etc. If not provided uses the colData slot from
the \code{SilacProteinExperiment} and \code{SilacPeptideExperiment}.}

\item{linkerDf}{A \code{data.frame} output from \code{\link{buildLinkerDf}}.}

\item{metadata}{A \code{list} to store any kind of experiment-wide
data; like authors, dates, machines used... If not provided uses the metadata
from the \code{SilacProteinExperiment} and \code{SilacPeptideExperiment}.}

\item{idColProt}{A \code{character} indicating which column from the
rowData (protein) should be used as ids. Should be the same used in
\code{\link{buildLinkerDf}}.}

\item{idColPept}{A \code{character} indicating which column from the
rowData (peptide) should be used as ids. Should be the same used in
\code{\link{buildLinkerDf}}.}

\item{linkedSubset}{A \code{logical} if subsetting should be linked between
proteins and peptide.}

\item{subsetMode}{A \code{character}, either 'protein' or 'peptide'
indicating which level should be used first when subsetting.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{timeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment timepoints.}

\item{proteinCol}{A \code{character}, which indicates the column name
in rowData(x) that defines to which protein a peptide is assigned.}
}
\value{
An object of class \code{SilacProteomicsExperiment}.
}
\description{
Constructor function for the ProteomicsExperiment class object.
It requires at minimum a \code{SilacProteinExperiment} and a
\code{SilacPeptideExperiment}. If the colData, metadata and metaoptions have
been already defined in those it is not necessary to give them again.
}
\section{Class description}{

See \link{SilacProteomicsExperiment-class} for details.
}

\section{Accessors}{

See \link{SilacProteomicsExperiment-accessors} for details.
}

\examples{

## assays
assays_protein <- list(expression = matrix(1:9, ncol = 3))

## colData
colData <- data.frame(sample = c('A1', 'A2', 'A3'),
                      condition = c('A', 'A', 'A'),
                    time = c(1, 2, 3))
## rowData
rowData_protein <- data.frame(prot_id = LETTERS[1:3])

## construct the ProteinExperiment
protExp <- SilacProteinExperiment(assays = assays_protein,
                                  rowData = rowData_protein,
                                  colData = colData,
                                  conditionCol = 'condition',
                                  timeCol = 'time')

## assays
assays_peptide <- list(expression = matrix(1:15, ncol = 3))

## colData
colData <- data.frame(sample = c('A1', 'A2', 'A3'),
                      condition = c('A', 'A', 'A'),
                      time = c(1, 2, 3))
## rowData
rowData_peptide <- data.frame(pept_id = letters[1:5],
                              prot_id = c('A', 'A', 'B', 'C', 'C'))
## construct the ProteinExperiment
peptExp <- SilacPeptideExperiment(assays = assays_peptide,
                                  rowData = rowData_peptide,
                                  colData = colData,
                                  conditionCol = 'condition',
                                  timeCol = 'time')

## list with the relationships
protein_to_peptide <- list(A = c('a', 'b'), B = c('c'), C = c('d', 'e'))
## function to build the data.frame
linkerDf <- buildLinkerDf(protIDs = LETTERS[1:3],
                          pepIDs  = letters[1:5],
                          protToPep = protein_to_peptide)

ProteomicsExp <- SilacProteomicsExperiment(SilacProteinExperiment = protExp,
                                           SilacPeptideExperiment = peptExp,
                                           linkerDf = linkerDf)

}
