% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetTimeCoverage.R
\docType{methods}
\name{upsetTimeCoverage}
\alias{upsetTimeCoverage}
\alias{upsetTimeCoverage,ProteinExperiment-method}
\alias{upsetTimeCoverage,PeptideExperiment-method}
\alias{upsetTimeCoverage,ProteomicsExperiment-method}
\title{Number of detected features per sample}
\usage{
upsetTimeCoverage(x, ...)

\S4method{upsetTimeCoverage}{ProteinExperiment}(x, assayName, conditionCol,
  maxMissing = 0, returnList = FALSE, ...)

\S4method{upsetTimeCoverage}{PeptideExperiment}(x, assayName,
  maxMissing = 0, conditionCol, returnList = FALSE, ...)

\S4method{upsetTimeCoverage}{ProteomicsExperiment}(x, assayName,
  maxMissing = 0, conditionCol, returnList = FALSE, ...)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{...}{Further arguments passed to \code{upset()}.}

\item{assayName}{Name of the assay to use in the plot.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{maxMissing}{A \code{numerical} indicating how many timepoints can a
protein/peptide miss.}

\item{returnList}{A \code{logical} indicating if the \code{list}
used for the plot should be returned instead.}
}
\value{
A barplot or a \code{data.frame}.
}
\description{
How many proteins/peptides are detected in each sample. Anything
else than NA is considered detected.
}
\examples{
data('wormsPE')
upsetTimeCoverage(x = ProtExp(wormsPE),
                assayName = 'ratio',
                maxMissing = 2)

}
