% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wormsPE}
\alias{wormsPE}
\title{ProteomicsExperiment with pulsed silac data from C. elegans strains}
\format{A \code{ProteomicsExperiment} object with 250 proteins and 3574
peptides in a total of 14 samples.
\describe{
  \item{colData}{A \code{DataFrame} with the design of the experiment:
  samples, timepoints, replicates...}
  \item{assaysProt}{A list of matrices with quantification data at protein
  level: total intensity (int_total), light isotope intensity (int_light),
  heavy isotope intensity (int_heavy) and heavy/light isotope intensty ratio
  (ratio).}
  \item{rowDataProt}{A \code{DataFrame} with 22 columns that contains general
  protein information: ids, gene names, molecular weight...}
  \item{assaysPep}{A list of matrices with quantification data at peptide
  level: total intensity (int_total), light isotope intensity (int_light),
  heavy isotope intensity (int_heavy) and heavy/light isotope intensty ratio
  (ratio).}
  \item{rowDataPept}{A \code{DataFrame} with 46 columns that contains general
  protein information: ids, amino acids counts, length...}
  \item{linkerDf}{A \code{data.frame} with 3574 rows and 4 columns. It
  contains the relationships between proteins and peptides in the
  ProteomicsExperiment object.}
}}
\usage{
wormsPE
}
\description{
A pre-built \code{ProteomicsExperiment} object with data from a pulsed silac
experiment done in \emph{C. elegans} by Visscher et al. 2016. It only
contains the data from the first 250 priteins and two old worms strains
(OW40 and OW450).
}
\details{
It is used as example in the pulsed silac vignette to illustrate the main
data analysis functions and in the examples of the documentation.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/28679685}
}
\keyword{datasets}
