% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAIC.R
\name{calculateAIC}
\alias{calculateAIC}
\title{Calculates the Akaike Information Criteria (AIC)}
\usage{
calculateAIC(modelList, smallSampleSize = TRUE)
}
\arguments{
\item{modelList}{a \code{list} with the model metrics,
the output from \code{\link{modelTurnover}}.}

\item{smallSampleSize}{a \code{logical} indicating if the AIC small
sample size correction formula should be used.}
}
\value{
a \code{list} with the model metrics (the given input) plus a matrix
named "AIC" with the AIC for each value
}
\description{
Calculates the AIC for each of the computed models. Requires that
\code{modelTurnover} is run with \code{reuturnModel = TRUE}.
}
\details{
The following formulas are used to compute the AIC and AICc (small sample
size correction):
\deqn{AIC = 2k - 2ln(logLik)}
\deqn{AICc = AIC + \frac{2k(k + 1)}{n - k - 1}}
}
\examples{
data('wormsPE')
wormsPE <- calculateIsotopeFraction(wormsPE, ratioAssay = 'ratio')

modelList <- modelTurnover(x = wormsPE[1:10],
                           assayName = 'fraction',
                           formula = 'fraction ~ 1 - exp(-k*t)',
                           start = list(k = 0.02),
                           mode = 'protein',
                           robust = FALSE,
                           returnModel = TRUE)

modelList <- calculateAIC(modelList, smallSampleSize = TRUE)

}
\seealso{
\code{\link{compareAIC}},
         \code{\link{modelTurnover}}
}
