% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateIsotopeRatio.R
\docType{methods}
\name{calculateIsotopeRatio}
\alias{calculateIsotopeRatio}
\alias{calculateIsotopeRatio,SilacProteinExperiment-method}
\alias{calculateIsotopeRatio,SilacPeptideExperiment-method}
\alias{calculateIsotopeRatio,SilacProteomicsExperiment-method}
\title{Ratio calculation}
\usage{
calculateIsotopeRatio(x, newIsotopeAssay, oldIsotopeAssay, ...)

\S4method{calculateIsotopeRatio}{SilacProteinExperiment}(x,
  newIsotopeAssay, oldIsotopeAssay)

\S4method{calculateIsotopeRatio}{SilacPeptideExperiment}(x,
  newIsotopeAssay, oldIsotopeAssay)

\S4method{calculateIsotopeRatio}{SilacProteomicsExperiment}(x,
  newIsotopeAssay, oldIsotopeAssay)
}
\arguments{
\item{x}{A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object.}

\item{newIsotopeAssay}{A \code{character} indicating the assay name that
has the new isotope intensity data.}

\item{oldIsotopeAssay}{A \code{character} indicating the assay name that
has the old isotope intensity data.}

\item{...}{Unused.}
}
\value{
A \code{SilacProteinExperiment}, \code{SilacPeptideExperiment} or a
\code{SilacProteomicsExperiment} object with an added assay named 'ratio'.
}
\description{
Calculates the ratio between the new isotope and the old
isotope (new/old).
}
\examples{
data('wormsPE')
calculateIsotopeRatio(x = wormsPE,
                      newIsotopeAssay = 'int_heavy',
                      oldIsotopeAssay = 'int_light')

}
