% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostStable.R
\docType{methods}
\name{mostStable}
\alias{mostStable}
\alias{mostStable,ProteinExperiment-method}
\alias{mostStable,PeptideExperiment-method}
\alias{mostStable,ProteomicsExperiment-method}
\title{Most stable proteins/peptides}
\usage{
mostStable(x, ...)

\S4method{mostStable}{ProteinExperiment}(x, assayName, n, conditionCol)

\S4method{mostStable}{PeptideExperiment}(x, assayName, n, conditionCol)

\S4method{mostStable}{ProteomicsExperiment}(x, assayName, n, mode,
  conditionCol)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{...}{Unused.}

\item{assayName}{Name of the assay to use.}

\item{n}{A \code{numeric} indicating how many proteins should be returned.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{mode}{A \code{character} indicating which level of data to use,
either "protein" or "peptide". Only relevant for ProteomicsExperiment
inputs.}
}
\value{
A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object with the n most stable proteins/peptides.
}
\description{
Finds which are the most stable proteins/peptides across
the entire experiment. These proteins/peptides can be used to estimate
the cell growth of each condition.
}
\details{
Proteins/peptides that are not found in all timepoints and in all
conditions are not considered. The stability is based on ranking
heavy label incorporation for each timepoint; therefore, lower values
are correlated to higher stability.
}
\examples{
data('mefPE')
mostStable(mefPE, assayName = 'fraction', n = 50)

}
