% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeptideExperiment.R
\name{PeptideExperiment}
\alias{PeptideExperiment}
\title{PeptideExperiment constructor}
\usage{
PeptideExperiment(assays, rowData = NULL, colData = NULL,
  conditionCol = NA, timeCol = NA, replicateIntCol = NA,
  replicateTimeCol = NA, proteinCol = NA, metadata = NULL)
}
\arguments{
\item{assays}{A named \code{list} of matrices (assays) with peptide level
data.}

\item{rowData}{A \code{data.frame} with peptide feature data
like protein names, molecular weight, etc.}

\item{colData}{A \code{data.frame} with sample information like
conditions, replicates, etc.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{timeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment timepoints.}

\item{replicateIntCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different expression replicates.}

\item{replicateTimeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different tiempoint replicates.}

\item{proteinCol}{A \code{character}, which indicates the column name
in rowData(x) that defines to which protein a peptide is assigned.}

\item{metadata}{A \code{list} to store any kind of experiment-wide
data; like authors, dates, machines used...}
}
\value{
An object of class PeptideExperiment
}
\description{
Constructor function for the PeptideExperiment class object.
For more information about the class check
\link{PeptideExperiment-class}.
}
\examples{

## assays
assays_peptide <- list(expression = matrix(1:15, ncol = 3))

## colData
colData <- data.frame(sample = c('A1', 'A2', 'A3'),
                      condition = c('A', 'A', 'A'),
                      time = c(1, 2, 3))
## rowData
rowData_peptide <- data.frame(pept_id = letters[1:5],
                              prot_id = c('A', 'A', 'B', 'C', 'C'))
## construct the ProteinExperiment
peptExp <- PeptideExperiment(assays = assays_peptide,
                             rowData = rowData_peptide,
                             colData = colData,
                             conditionCol = 'condition',
                             timeCol = 'time')

}
