% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributionAssay.R
\docType{methods}
\name{plotDistributionAssay}
\alias{plotDistributionAssay}
\alias{plotDistributionAssay,ProteinExperiment-method}
\alias{plotDistributionAssay,PeptideExperiment-method}
\alias{plotDistributionAssay,ProteomicsExperiment-method}
\title{Distribution of assay data per condition and timepoint.}
\usage{
plotDistributionAssay(x, ...)

\S4method{plotDistributionAssay}{ProteinExperiment}(x, assayName,
  plotType = "boxplot", returnDataFrame = FALSE, conditionCol, timeCol)

\S4method{plotDistributionAssay}{PeptideExperiment}(x, assayName,
  plotType = "boxplot", returnDataFrame = FALSE, conditionCol, timeCol)

\S4method{plotDistributionAssay}{ProteomicsExperiment}(x, assayName,
  mode = "protein", plotType = "boxplot", returnDataFrame = FALSE,
  conditionCol, timeCol)
}
\arguments{
\item{x}{A \code{ProteinExperiment}, \code{PeptideExperiment} or a
\code{ProteomicsExperiment} object.}

\item{...}{Unused.}

\item{assayName}{Name of the assay to use in the plot.}

\item{plotType}{A \code{character} indicating which geometry to plot:
'boxplot' or 'density'. (default = 'density')}

\item{returnDataFrame}{A \code{logical} indicating if the \code{data.frame}
used for the plot should be returned instead.}

\item{conditionCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different experiment conditions.}

\item{timeCol}{A \code{character}, which indicates the column name
in colData(x) that defines the different timepoints.}

\item{mode}{A \code{character} indicating which level of data to use,
either "protein" or "peptide". Only relevant for ProteomicsExperiment
inputs.}
}
\value{
A ggplot2 object or a \code{data.frame} with the data that would be
plotted.
}
\description{
Plot the distribution of the data stored in an assay using
boxplots or density distributions.
}
\examples{
data('wormsPE')
plotDistributionAssay(wormsPE, assayName = 'ratio')

}
